+++ pages/_app.tsx
@@ -1,4 +1,3 @@
-import Gnb from "@/components/Gnb";
 import { globalStyles } from "@/styles/global";
 import { Global } from "@emotion/react";
 import type { AppProps } from "next/app";
