+++ pages/shops/[shopId]/notices/[noticeId]/components/PostDetail.tsx
@@ -1,92 +1,64 @@
-import getWageIncreaseText from "@/components/Post/utils/getWageIncreaseText";
-import { NoticeData } from "@/hooks/useNoticeData";
-import formatTimeRange from "@/lib/utils/formatTimeRange";
-import { body1Regular, body2Regular, h1Regular } from "@/styles/fontsStyle";
 import styled from "@emotion/styled";
-import Image from "next/image";
-import clockIcon from "@/public/images/clock-icon 1.svg";
-import locationIcon from "@/public/images/location.svg";
+import PostCard from "./PostCard";
+import { NoticeData } from "@/hooks/useNoticeData";
+import { body1Regular, h1Regular } from "@/styles/fontsStyle";
 
-export default function PostDetail({ noticeData }: { noticeData: NoticeData }) {
-  if (noticeData) {
-    const duration = formatTimeRange(
-      noticeData.item.startsAt,
-      noticeData.item.workhour,
-    );
-    const shopDescription = noticeData.item.shop.item.description;
-    const address = `${noticeData.item.shop.item.address1} ${noticeData.item.shop.item.address2}`;
-    const hourlyPay = noticeData.item.hourlyPay;
-    const originalHourlyPay = noticeData.item.shop.item.originalHourlyPay;
-    const wageIncreaseText = getWageIncreaseText(hourlyPay, originalHourlyPay);
-
-    return (
-      <Wrapper>
-        <Wage>시급</Wage>
-        <WageContainer>
-          <HourlyPay>{hourlyPay}원</HourlyPay>
-          <WageFlagStyle>{wageIncreaseText}</WageFlagStyle>
-        </WageContainer>
-        <Container>
-          <Image src={clockIcon} alt="clock_icon" />
-          <InfoStyle>{duration}</InfoStyle>
-        </Container>
-        <Container>
-          <Image src={locationIcon} alt="location_icon" />
-          <InfoStyle>{address}</InfoStyle>
-        </Container>
-        <DescriptionStyle>{shopDescription}</DescriptionStyle>
-      </Wrapper>
-    );
-  }
+interface PostDetailProps {
+  noticeData: NoticeData;
 }
 
-const Wrapper = styled.div`
-  margin-top: 16px;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  width: 100%;
-`;
-
-const Wage = styled.span`
-  ${body1Regular}
-  color: var(--The-julge-purple-40, #905CB9);
-`;
-
-const HourlyPay = styled.span`
-  ${h1Regular}
-`;
-
-const WageFlagStyle = styled.div`
-  display: flex;
-  height: 36px;
-  padding: 12px;
-  align-items: center;
-  gap: 6px;
-  width: 110px;
-
-  color: var(--The-julge-gray-00, #ffffff);
-  border-radius: 20px;
-  background: var(--The-julge-purple-40, #905cb9);
-  ${body2Regular}
-`;
+export default function PostDetail({ noticeData }: PostDetailProps) {
+  return (
+    <Container>
+      <ShopDetails>
+        <Category>{noticeData.item.shop.item.category}</Category>
+        <Title>{noticeData.item.shop.item.name}</Title>
+      </ShopDetails>
+      <PostCard noticeData={noticeData} />
+      <PostDescription>
+        <DescriptionHeading>공고 설명</DescriptionHeading>
+        <DescriptionText>{noticeData.item.description}</DescriptionText>
+      </PostDescription>
+    </Container>
+  );
+}
 
 const Container = styled.div`
   display: flex;
-  flex-direction: row;
-  gap: 6px;
+  flex-direction: column;
+  justify-content: flex-start;
+  padding: 60px 238px;
+  gap: 24px;
+  background: var(--The-julge-gray-05, #fafafa);
 `;
 
-const WageContainer = styled(Container)`
+const ShopDetails = styled.div`
+  display: flex;
+  flex-direction: column;
   gap: 8px;
 `;
 
-const InfoStyle = styled.span`
-  color: var(--The-julge-gray-50, #7d7986);
-  ${body2Regular}
+const Category = styled.span`
+  color: var(--The-julge-purple-40, #905cb9);
+  ${body1Regular}
 `;
 
-const DescriptionStyle = styled.span`
-  ${body2Regular}
+const Title = styled.span`
+  ${h1Regular}
 `;
+
+const PostDescription = styled.div`
+  display: flex;
+  padding: 32px;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 32px;
+  width: 100%;
+  border-radius: 12px;
+  background: var(--The-julge-gray-10, #f2f2f3);
+  ${body1Regular}
+`;
+
+const DescriptionHeading = styled.span``;
+
+const DescriptionText = styled.p``;
